requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
gunicorn==19.9.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
